requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
livekit-agents==0.5.1
livekit-api
python-dotenv==1.0.0
uuid